requirements.txt
click==8.1.7
dnspython==2.6.1
fire==0.6.0
pymongo==4.8.0
python-dotenv==1.0.1
randomname==0.2.1
six==1.16.0
termcolor==2.4.0
